Rename school factory file
