+++ apps/server/src/modules/provisioning/strategy/iserv/iserv.strategy.spec.ts
@@ -3,8 +3,7 @@ import { ObjectId } from '@mikro-orm/mongodb';
 import { Test, TestingModule } from '@nestjs/testing';
 import { LegacySchoolDo, RoleName, User, UserDO } from '@shared/domain';
 import { SystemProvisioningStrategy } from '@shared/domain/interface/system-provisioning.strategy';
-import { schoolFactory, setupEntities, userDoFactory, userFactory } from '@shared/testing';
-import { legacySchoolDoFactory } from '@shared/testing/factory/domainobject/school.factory';
+import { legacySchoolDoFactory, schoolFactory, setupEntities, userDoFactory, userFactory } from '@shared/testing';
 import { OAuthSSOError } from '@src/modules/oauth/error/oauth-sso.error';
 import { LegacySchoolService } from '@src/modules/school';
 import { UserService } from '@src/modules/user';
