+++ apps/server/src/modules/user-login-migration/service/school-migration.service.spec.ts
@@ -5,8 +5,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { ValidationError } from '@shared/common';
 import { LegacySchoolDo, Page, UserDO, UserLoginMigrationDO } from '@shared/domain';
 import { UserLoginMigrationRepo } from '@shared/repo/userloginmigration/user-login-migration.repo';
-import { setupEntities, userDoFactory, userLoginMigrationDOFactory } from '@shared/testing';
-import { legacySchoolDoFactory } from '@shared/testing/factory/domainobject/school.factory';
+import { legacySchoolDoFactory, setupEntities, userDoFactory, userLoginMigrationDOFactory } from '@shared/testing';
 import { LegacyLogger } from '@src/core/logger';
 import { ICurrentUser } from '@src/modules/authentication';
 import { LegacySchoolService } from '@src/modules/school';
