+++ apps/server/src/shared/testing/factory/domainobject/index.ts
@@ -1,7 +1,7 @@
 export * from './board';
 export * from './tool';
 export * from './do-base.factory';
-export * from './school.factory';
+export * from './legacy-school.factory';
 export * from './user-login-migration-do.factory';
 export * from './lti-tool.factory';
 export * from './pseudonym.factory';
